+++ backend/etc/crate_setup.sql
@@ -1,1 +1,2 @@
 create table blogpost (id string primary key, created timestamp, text string, creator string )
+create table users (id string primary key, name string, password string )